move zoom_in to tools/
